pyproject.toml
[project]
name = "ai-research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "gradio>=6.6.0",
    "openai>=2.21.0",
    "pypdf>=6.7.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "youtube-transcript-api>=1.2.4",
]
